refactor go packages
