+++ .release/release-metadata.hcl
@@ -1,6 +1,6 @@
 url_docker_registry_dockerhub = "https://hub.docker.com/r/hashicorp/consul-lambda-registrator"
 url_docker_registry_ecr       = "https://gallery.ecr.aws/hashicorp/consul-lambda-registrator"
-url_source_repository         = "https://github/hashicorp/terraform-aws-consul-lambda-registrator"
+url_source_repository         = "https://github/hashicorp/terraform-aws-consul-lambda"
 url_project_website           = "https://www.consul.io/docs/lambda"
-url_license                   = "https://github.com/hashicorp/terraform-aws-consul-lambda-registrator/blob/main/LICENSE"
+url_license                   = "https://github.com/hashicorp/terraform-aws-consul-lambda/blob/main/LICENSE"
 url_release_notes             = "https://www.consul.io/docs/release-notes"
