+++ consul-lambda/.golangci.yml
