+++ consul-lambda/build-scripts/version.sh
