+++ consul-lambda/client/ssm.go
