+++ consul-lambda/consul-lambda-extension/lambda_client.go
@@ -1,4 +1,4 @@
-package extension
+package main
 
 import (
 	"bytes"
