+++ consul-lambda/consul-lambda-registrator/.dockerignore
