+++ consul-lambda/consul-lambda-registrator/Dockerfile
@@ -20,7 +20,7 @@ LABEL description="consul-lambda-registrator automatically registers Lambda func
       org.opencontainers.image.authors="Consul Team <consul@hashicorp.com>" \
       org.opencontainers.image.description="consul-lambda-registrator automatically registers Lambda functions into Consul." \
       org.opencontainers.image.documentation="https://www.consul.io/docs/lambda" \
-      org.opencontainers.image.source="https://github.com/hashicorp/terraform-aws-consul-lambda-registrator" \
+      org.opencontainers.image.source="https://github.com/hashicorp/terraform-aws-consul-lambda" \
       org.opencontainers.image.title=$BIN_NAME \
       org.opencontainers.image.url="https://www.consul.io/" \
       org.opencontainers.image.vendor="HashiCorp" \
