+++ consul-lambda/consul-lambda-registrator/Makefile
