+++ consul-lambda/consul-lambda-registrator/environment_test.go
