+++ consul-lambda/consul-lambda-registrator/fixtures/create_function.json
