+++ consul-lambda/consul-lambda-registrator/fixtures/tag_resource.json
