+++ consul-lambda/consul-lambda-registrator/fixtures/unsupported.json
