+++ consul-lambda/consul-lambda-registrator/fixtures/untag_resource.json
