+++ consul-lambda/consul-lambda-registrator/lambda_client.go
