+++ consul-lambda/consul-lambda-registrator/main.go
