+++ consul-lambda/proxy/conn.go
