+++ consul-lambda/proxy/conn_test.go
