+++ consul-lambda/proxy/proxy.go
