+++ consul-lambda/proxy/testdata/server-cert.pem
