+++ consul-lambda/structs/service.go
