+++ consul-lambda/trace/logger.go
