+++ consul-lambda/trace/trace.go
